feat(pagination+pagination-nav): remove need for custom active focus style (#1384)
* [pagination-mixin] use btn-primary to get focus styling

Avoids the need for custom CSS until properly addressed in Bootstrap V4.

See https://github.com/twbs/bootstrap/issues/24838

* Update pagination-nav.js

* Delete pagination-nav.vue

* Update pagination.js

* Delete pagination.vue

* ESLint

* ESLint

* Include active style when pagination disabled

* Update pagination.js
